third_party/WebKit/LayoutTests/fast/dom/shadow/get-element-by-id-in-shadow-mutation-expected.txt

CONSOLE WARNING: line 12: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests to ensure that ShadowRoot.getElementById works even after mutation

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS null is shadow.getElementById('foo')
PASS e0a is shadow.getElementById('foo')
PASS e0b is shadow.getElementById('foo')
PASS e0c is shadow.getElementById('foo')
PASS e0b is shadow.getElementById('foo')
PASS e0a is shadow.getElementById('foo')
PASS null is shadow.getElementById('foo')
PASS successfullyParsed is true

TEST COMPLETE
